# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API sistem absensi mahasiswa berbasis geolocation. Semua respons memakai envelope {success, data, message, errors}.",
        "title": "Absensi Mahasiswa API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/locations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin: Locations"
                ],
                "summary": "Daftar lokasi",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Envelope"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin: Locations"
                ],
                "summary": "Buat lokasi",
                "parameters": [
                    {
                        "description": "Lokasi",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LocationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.Envelope"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.Envelope"
                        }
                    }
                }
            }
        },
        "/locations/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin: Locations"
                ],
                "summary": "Ubah lokasi",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID lokasi",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Lokasi",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LocationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Envelope"
                        }
                    }
                }
            }
        },
...
        "/reports/attendance": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin: Reports"
                ],
                "summary": "Laporan absensi (paginated)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "YYYY-MM-DD",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "YYYY-MM-DD",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ID jadwal",
                        "name": "schedule_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Halaman",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Per halaman (1-100)",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Envelope"
                        }
                    }
                }
            }
        },
        "/reports/attendance/export": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
                ],
                "tags": [
                    "Admin: Reports"
                ],
                "summary": "Export laporan absensi ke Excel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "YYYY-MM-DD",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "YYYY-MM-DD",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ID jadwal",
                        "name": "schedule_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.LocationRequest": {
            "type": "object",
            "required": [
                "latitude",
                "longitude",
                "name",
                "radius"
            ],
            "properties": {
                "latitude": {
                    "type": "number",
                    "maximum": 90,
                    "minimum": -90
                },
                "longitude": {
                    "type": "number",
                    "maximum": 180,
                    "minimum": -180
                },
                "name": {
                    "type": "string"
                },
                "radius": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Format: \"Bearer {access_token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}